requirements.txt
fastapi == 0.68.1
uvicorn == 0.15.0
aioredis == 2.0.0
Jinja2 == 3.0.2
python-multipart == 0.0.5
asyncpg == 0.25.0
python-dotenv == 0.19.2
python-jose == 3.3.0
httpx == 0.21.1
python-decouple == 3.5
tomlkit == 0.8.0
